+++ packages/sshnp_gui/lib/src/presentation/screens/new_connection_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:sshnp_gui/src/presentation/widgets/profile/new_profile_form.dart';
+import 'package:sshnp_gui/src/presentation/widgets/profile/profile_form.dart';
 
 import '../../utils/sizes.dart';
 import '../widgets/navigation/app_navigation_rail.dart';
